UefiCpuPkg/MpInitLib: Always use AP Create protocol if PcdSevSnpApicIds is set
Currently, the first time an AP is started for an SEV-SNP guest, it relies
on the VMSA as set by the hypervisor. If the list of APIC IDs has been
retrieved, this is not necessary. Instead, use the SEV-SNP AP Create
protocol to start the AP for the first time and thereafter.

To use the AP Create protocol, the page used to hold the VM state must be
updated to set the VMSA attribute. The RMPADJUST instruction is used to do
this, but the VMSA attribute is only set when running at VMPL0. When an
SVSM is present, use the SVSM_CORE_CREATE_VCPU and SVSM_CORE_DELTE_VCPU
calls to add or remove the VMSA attribute on a page instead of issuing
the RMPADJUST instruction directly.

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>