Ovmfpkg/VmgExitLib: Extend VmgExitLib to handle SVSM related services
Add initial support for the new VmgExitLib interfaces to the OvmfPkg
version of the library. The initial implementation will fully implement
the SVSM presence check API, with later patches fully implementing the
other interfaces.

The SVSM presence check, VmgExitSvsmPresent(), determines the presence
of an SVSM by checking if an SVSM has been advertised in the SEV-SNP
Secrets Page.

Rebase (osteffen):
- Vgm is now called CC
- TODO: rename VgmExitSvsm.*

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>