OvmfPkg: Create a calling area used to communicate with the SVSM
An SVSM requires a calling area page whose address (CAA) is used by the
SVSM to communicate and process the SVSM request.

Add a pre-defined page area to the OvmfPkg and define corresponding PCDs
used to communicate the location and size of the area.

Rebase: (Osteffen):
resolved conflict in PCD ids!

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>